requirements.txt
argparse~=1.4.0
pcpp~=1.21
pycparser~=2.19
pyelftools~=0.26